pyproject.toml
[project]
name = "tg-stats"
version = "0.1.0"
description = "Telegram analytics bot with multi-group support"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[all]>=21.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "structlog>=23.0",
    "emoji>=2.8",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
    "black>=23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["tgstats*"]
exclude = ["migrations*", "tests*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.black]
line-length = 100
target-version = ["py312"]
extend-exclude = "migrations"

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = ["migrations", ".venv", "venv"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
